pkg/cmd/server: move ClientCARegistrationHook setup to kube master_config.go
